Rename grid features file
