+++ src/web/openlayers/features/survey/grid-features.ts
